+++ examples/basic_usage.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from corenorm import CoreNorm
 
 X = np.array([[1.0, 10.0], [2.0, 11.0], [3.0, 12.0], [4.0, 13.0], [40.0, 60.0]])