+++ examples/pipeline.py
@@ -2,6 +2,7 @@ from sklearn.datasets import load_diabetes
 from sklearn.linear_model import Ridge
 from sklearn.model_selection import train_test_split
 from sklearn.pipeline import make_pipeline
+
 from corenorm import CoreNorm
 
 X, y = load_diabetes(return_X_y=True)