+++ examples/save_restore.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from corenorm import CoreNorm
 
 X = np.array([[1.0, 2.0], [2.0, 3.0], [4.0, 100.0]])